Add shared settings module and persist online trainer params
